better static
